+++ HISTORY.rst
@@ -16,6 +16,7 @@ Unreleased
 
 * ALCN-1008: Raise an error when a user try to "deliver" a round with partially available pickings
 * Add user authentication to webservices from connector_esb ALCN-1044
+* ALCN-898: Add the module sale_cancel_remaining
 
 **Bugfixes**